v1.0.0 of Weather Metrics
I have written all the code for the front-end up. This morning I implemented a Flask server and that is working now. All is left to do is fix up the navigation into a settings cog, and call the data from our API to be displayed on the front end.